Consolidate rootservlet into listbookservlet Update README Simplify web.xml
